ViewBindings
- move bindings into own file
